Finally got a working NoteReader.
I finally built a working NoteReader class. I have also modified the
Note class. It now overrides the equals() and hashCode() method of the
Object class. This will have to be cherry-picked to the rest of the
branches as well. Hojjiin Itti Fufa

Minor disturbance, I accidentally delete the NoteReader class so I had
to put it back. It tooke while, but finally fixed it.