Built a full NoteReader class
The NoteReader class has been completed according to the functional specifications. It now has a method to load from the primary file and the backup file. I've extracted the reading operation into it's own method because it's basically similar between the load() and loadBackup() methods with the difference being that the loadBackup() uses a different FileInputStream than the load(). Hojjiin Itti Fufa.