Completed the WordWrap inner class
I've completed the WordWrap inner class with it's corresponding wrap()
method. It take two parameters, the text to be wrapped and the
lineLenght to wrap at. And i've removed the alignment menu
Item. I've also minimized the width the scrollpane can take and set menu
items for each line length that the wrap() method will use.