updated the removeNote() and counter
I've updated the removeNote() method to deduct from the counter of the
AxeereraaUI. If the last visible Note is deleted then the application
will close itself.
I've also renamed the counter instance variable to COUNTER b/c it's a
static variable available to the entire class not to an object. It will
now be accessed via the class reference.