Has done a very big one
I've added the flexmark-java library to the bulid.gradle file so as to parse the written markdown.
I've also change the way the "about" and "contact developer" menu items display their information.
I've added a new class DisplayMarkdown that will handle the parsing and
displaying of the markdown. It's static methods convertToMarkdown() and
displayMarkdown() are responsible for parsing and displaying using a
JEditorPane respectively.
I've made a lot of improvements to the AxeereraaUI class and it's
methods, to name a few, changed the preview menu item into a menu and
added "showMarkdown" and "showRawText" as menu options. They each call
their respective methods showMarkdown() and showRawText(). I've also
added the displayDialog() method to handle the dialog creation for the
"about" & "contact developer" menu items.
Any changes I've forgotten that were not added to this commit message
can be infered from the source code. Hojjin Itti Fufa.