Finally got it to work as a JAR
I've finally got it to work as a JAR. I used gradle to do it instead of
the Intellij-IDEA's own artifacts tool (why b/c it doesn't fucking
work). I've set the GUI designer to generate the source code of the UI
and thus it works. All the code modifications are self explanatory with
the exception of the build.gradle file.
I've added a new directory by the name of libs to contain the
form_rt.jar file from the Intellij-IDEA's form generator and compiled it
in the dependecies section. And now we release.