Improved the way font is loaded
I've modified & improved the way the font is loaded, it now uses creates
an new InputStream according to the file source, creates a Font object
from that and drives a new Font object with a modified style & size.
There is still the issue of it not applying the Font object on the first
instance of the application.